move creative tabs event out of client code
